\name{fiveGyMAS5}
\docType{data}
\alias{fiveGyMAS5}
\title{Example data for the rHVDM package (time course of irradiated human T cells)}
\description{
  fiveGyMAS5: the data set proper is an Expression set (Biobase). MOLT4 p53-wild type human T-cells irradiated
  @ 5 Gy, 7 time points. Triplicate experiment.
}
\format{R data (.rda)}
\keyword{datasets}
