\name{HVDMreport}

\alias{HVDMreport}

\title{generates an HTML report using the input}
\description{
  This method generates report using one of the HVDM outputs (either from the \code{training()}, \code{screeening()} or \code{fitgene()} commands.)
}
\usage{
HVDMreport(HVDM,name)
}
\arguments{
  \item{HVDM}{ a list output either by a \code{training()}, \code{screening()} or \code{fitgene()} command.}
	\item{name}{an optional argument allowing to specify the name of the HTML file (the .HTML extension is added automatically).
	}
}
\details{ 
	If the output has been generated by \code{training()} or \code{fitgene()}, the HTML report contains
	four sections:
	
	1) Functions inputs.
	
	2) Model score.
	
	3) Fitted parameters.
	
	4) Visual comparison of model and data.
	
	In the \code{fitgene()} case comparisons with correponding results for genes in the training set are given.
	
	If the output has been generated by the \code{screening()} command, the HTML report contains the following sections:

	1) Inputs to the preceding \code{training()} command.
	
	2) Inputs to the \code{screening()} command (such as bounds used to classify genes).
	
	3) Results.
	
	4) List of putative targets of the transcription factor under review.
	
	The HTML report is genereated in the current working directory. Graphic files are stored in a subdirectory with a similar name.
	
}

\value{
	Nothing is returned, but a message indicating the name of the HTML file and its location on the hard drive is given.
}
\references{ 
M. Barenco, D. Tomescu, D. Brewer, R. Callard, J. Stark, M. Hubank (2006) Ranked predictions
of p53 targets using Hidden Variable Dynamic Modelling. \emph{Genome Biology}, \bold{V7(3)}, R25.
}
\author{ Martino Barenco }

\note{ The HTML report is generated in the working directory.
}

\seealso{\code{\link{training}},\code{\link{screening}},\code{\link{fitgene}}}

\examples{
data(HVDMexample)
tHVDMp53<-training(eset=fiveGyMAS5,genes=p53traingenes,degrate=0.8,actname="p53")
#HVDMreport(tHVDMp53)
#HVDMreport(tHVDMp53, name="p53training")
}
\keyword{ IO }
\keyword{ ts }
