\name{rHVDMplattforms}
\docType{data}
\alias{rHVDMplattforms}
\title{plattform-specific repository allowing to estimate measurement errors}
\description{
	plattform-specific repository allowing to estimate measurement errors, 
	plattforms covered:
	1 - Affymetrix Human genome 133A 3-prime array  
}
\format{R data (.rda)}
\keyword{datasets}
