\name{twodosesMAS5}
\docType{data}
\alias{twodosesMAS5}
\title{Example data for the rHVDM package (time course of irradiated human T cells, two different doses)}
\description{
  fiveGyMAS5: the data set proper is an Expression set (Biobase). MOLT4 p53-wild type human T-cells irradiated
  @ 5 Gy, 7 time points. Triplicate experiment. Another single time course, same cells and conditions @ 0.5 Gy, with 2 extra points (16 and 20 hours after irradiation)
}
\format{R data (.rda)}
\keyword{datasets}
